+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/BooleanExtensions.kt
@@ -58,30 +58,4 @@ inline fun Boolean.toFloat():Float = if(this) 1f else 0f
 
 /**将当前布尔值转化为双精度浮点数类型的1或0。*/
 inline fun Boolean.toDouble():Double = if(this) 1.0 else 0.0
-
-
-///**转化为指定的相反数。默认为1或-1。*/
-//inline fun Boolean.toOpsInt(value: Int = 1): Int = if(this) value else -value
-//
-///**转化为指定的相反数。默认为1或-1。*/
-//inline fun Boolean.toOpsLong(value: Long = 1): Long = if(this) value else -value
-//
-///**转化为指定的相反数。默认为1或-1。*/
-//inline fun Boolean.toOpsFloat(value: Float = 1f): Float = if(this) value else -value
-//
-///**转化为指定的相反数。默认为1或-1。*/
-//inline fun Boolean.toOpsDouble(value: Double = 1.0): Double = if(this) value else -value
-//
-//
-///**转化为指定的相反数。默认为1或-1。如果为null，则转化为0。*/
-//inline fun Boolean?.toOpsInt(value: Int = 1): Int = this?.toOpsInt(value) ?: 0
-//
-///**转化为指定的相反数。默认为1或-1。如果为null，则转化为0。*/
-//inline fun Boolean?.toOpsLong(value: Long = 1): Long = this?.toOpsLong(value) ?: 0
-//
-///**转化为指定的相反数。默认为1或-1。如果为null，则转化为0。*/
-//inline fun Boolean?.toOpsFloat(value: Float = 1f): Float = this?.toOpsFloat(value) ?: 0f
-//
-///**转化为指定的相反数。默认为1或-1。如果为null，则转化为0。*/
-//inline fun Boolean?.toOpsDouble(value: Double = 1.0): Double = this?.toOpsDouble(value) ?: 0.0
 //endregion