+++ breeze-dsl/src/main/kotlin/com/windea/breezeframework/dsl/Dsl.kt
@@ -51,6 +51,12 @@ annotation class InlineDslFunction
 @Target(AnnotationTarget.FUNCTION)
 annotation class DslFunction
 
+//TODO 让注解处理器生效
+/**注明了这个注解定义了一个领域特定语言文件。这个文件是由Kotlin脚本编写的，其输出结构是一个[DslDocument]。*/
+@MustBeDocumented
+@Target(AnnotationTarget.FILE)
+annotation class DslFile
+
 /**Dsl文档。*/
 interface DslDocument {
 	override fun toString():String